Moved progresss tracking to core
