+++ kolibri/core/assets/src/core-store.js
@@ -11,6 +11,10 @@ const initialState = {
     login_modal_state: false,
     is_admin_or_superuser: false,
     fullname: '',
+    logging: {
+      summary: { progress: 0 },
+      session: { },
+    },
   },
 };
 
@@ -40,6 +44,35 @@ const mutations = {
   },
   SET_ADMIN_STATUS(state, value) {
     state.core.is_admin_or_superuser = value;
+  },
+  SET_LOGGING_SUMMARY_STATE(state, summaryState) {
+    state.core.logging.summary = summaryState;
+  },
+  SET_LOGGING_SUMMARY_ID(state, summaryId) {
+    state.core.logging.summary.id = summaryId;
+  },
+  SET_LOGGING_SESSION_ID(state, sessionId) {
+    state.core.logging.session.id = sessionId;
+  },
+  SET_LOGGING_SESSION_STATE(state, sessionState) {
+    state.core.logging.session = sessionState;
+  },
+  SET_LOGGING_PROGRESS(state, sessionProgress, summaryProgress) {
+    state.core.logging.session.progress = sessionProgress;
+    state.core.logging.summary.progress = summaryProgress;
+  },
+  SET_LOGGING_COMPLETION_TIME(state, time) {
+    state.core.logging.summary.completion_timestamp = time;
+  },
+  SET_LOGGING_TIME(state, sessionTime, summaryTime, currentTime) {
+    state.core.logging.session.end_timestamp = currentTime;
+    state.core.logging.summary.end_timestamp = currentTime;
+    state.core.logging.session.time_spent = sessionTime;
+    state.core.logging.summary.time_spent = summaryTime;
+  },
+  SET_LOGGING_PENDING(state, summaryPending, sessionPending) {
+    state.core.logging.summary.pending_create = summaryPending;
+    state.core.logging.session.pending_create = sessionPending;
   },
 };
